+++ src/main/java/com/hackathon/project/domain/Roadmap/dto/RoadmapCreateRequestDTO.java
@@ -19,7 +19,7 @@ public class RoadmapCreateRequestDTO {
     public static class Transcript {
 
         private List<Course> courses;
-        private int totalCredits;
+        private double totalCredits;
         private int totalMajorCredits;
         private int totalGeneralCredits;
         private double averageGPA;
@@ -37,7 +37,7 @@ public class RoadmapCreateRequestDTO {
         private String courseType;       // 이수구분
         private String teachingArea;     // 교직영역
         private String selectedArea;     // 선택영역
-        private int credits;             // 학점
+        private double credits;             // 학점
         private String evaluationType;   // 평가방식
         private String grade;             // 등급
         private double gradePoint;        // 평점
